# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A real-time chat server with JWT authentication, channel management, and WebSocket support",
        "title": "Go Chat API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://github.com/your-org/go-chat",
            "email": "support@gochat.dev"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:9876",
    "basePath": "/",
    "paths": {
        "/api/audit": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Get audit logs with optional filtering (system admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audit Logs"
                ],
                "summary": "Get audit logs with filtering",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by channel ID",
                        "name": "channel_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by actor ID",
                        "name": "actor_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by action type",
                        "name": "action",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of results per page (default: 20, max: 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Audit logs retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/internal_api.AuditLogsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Admin access required",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/channels/{id}/audit": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Get audit logs for a specific channel (only channel owners can view)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audit Logs"
                ],
                "summary": "Get channel audit logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Channel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of results per page (default: 20, max: 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Audit logs retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/internal_api.AuditLogsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Only channel owners can view audit logs",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Channel not found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal_api.AuditLogResponse": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string",
                    "example": "BAN_USER"
                },
                "actor": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "string",
                            "example": "abc12345"
                        },
                        "username": {
                            "type": "string",
                            "example": "admin_user"
                        }
                    }
                },
                "actor_id": {
                    "type": "string",
                    "example": "abc12345"
                },
                "channel": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "string",
                            "example": "xyz123"
                        },
                        "name": {
                            "type": "string",
                            "example": "general"
                        }
                    }
                },
                "channel_id": {
                    "type": "string",
                    "example": "xyz123"
                },
                "created_at": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Banned user from channel"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": true
                },
                "target": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "string",
                            "example": "def67890"
                        },
                        "username": {
                            "type": "string",
                            "example": "banned_user"
                        }
                    }
                },
                "target_id": {
                    "type": "string",
                    "example": "def67890"
                }
            }
        },
        "internal_api.AuditLogsResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api.AuditLogResponse"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "internal_api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "username cannot be empty"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CookieAuth": {
            "description": "JWT token stored in HTTP-only cookie",
            "type": "apiKey",
            "name": "token",
            "in": "cookie"
        }
    }
}